third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-hierarchy-exception-expected.txt

Appending an ancestor to a shadow tree should throw an exception

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowDiv.appendChild(container) threw exception HierarchyRequestError: A Node was inserted somewhere it doesn't belong..
PASS successfullyParsed is true

TEST COMPLETE
